screen: review polish for the PWL sense one-pole (comments and tests)
Review follow-ups to 60bd96a, no behaviour change:
- pwl_sense_alpha_'s comment now says what it is (the one-pole's
  per-sample charge fraction, 1 - exp(-1/(tau * fs))), not a mislabelled
  tau-in-samples.
- pwl_sense_'s double rationale reworded: the drive/sense/peak path is
  already double and the state accumulates across the line's samples,
  rather than implying every per-sample IIR is double.
- The start_line reset comment carries its load-bearing half: the reset
  also keeps the sensor stateless across lines, so the limiter loop
  stays first-order and the sense lag can't cause hunting.
- The spike-test lambda uses the same explicit static_cast<float> the
  reworked feed_lines does.
- render.md notes a dense sub-microsecond oscillation still registers
  at its mean - only the AC part is stripped.

Rebuilt, ctest green; the pattern.bin motivating-case frames are
byte-identical to the previous run (R 201.3 beside the blocks vs 200.9
above).

+++ lib/test/VideoTest.cpp
@@ -827,7 +827,8 @@ TEST_CASE("peak-white limiter: sub-microsecond ringing is ignored, a broad overl
       for (std::size_t k = 0; k < kActive; ++k) {
         const bool spiking = k >= kActive / 2 && k < kActive / 2 + spike_samples;
         pic.push_back(video::ChromaSample{.luma = spiking ? -0.7f : -0.15f});
-        hbeam.push_back(video::BeamSample{.h_phase = 0.2f + 0.7f * static_cast<float>(k) / kActive});
+        hbeam.push_back(
+            video::BeamSample{.h_phase = 0.2f + 0.7f * static_cast<float>(k) / static_cast<float>(kActive)});
         vbeam.push_back(video::VSample{.v_phase = 0.5f});
       }
       screen.process(pic, hbeam, vbeam);
